add build_in_vm.ps1 and fix parallel build

+++ scripts/build_inside.ps1
@@ -1,21 +1,29 @@
 Write-Host "====== Make osgeo4w links..."
 # Due to some cmake errors, the relative path is lost and resources are searched starting
 # at '/'. These links helps cmake to find the needed resources
-New-Item -Path C:\bin -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\bin
-New-Item -Path C:\lib -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\lib
-New-Item -Path C:\include -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\include
+if (-not (Test-Path -Path C:\bin)) {
+  New-Item -Path C:\bin -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\bin
+}
+if (-not (Test-Path -Path C:\lib)) {
+  New-Item -Path C:\lib -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\lib
+}
+if (-not (Test-Path -Path C:\include)) {
+  New-Item -Path C:\include -ItemType SymbolicLink -Value $Env:OSGEO4W_ROOT\include
+}
 
 Write-Host "====== Copy plugin sources to QGIS project..."
-New-Item -ItemType Directory -Force -Path QGIS/src/plugins/primitive_editing
-Copy-Item -Path "CMakeLists_inside.txt" -Destination "QGIS/src/plugins/primitive_editing/CMakeLists.txt" -Force
-Copy-Item -Path "src" -Destination "QGIS/src/plugins/primitive_editing/" -Recurse -Force
-Copy-Item -Path "resources" -Destination "QGIS/src/plugins/primitive_editing/" -Recurse -Force
-Copy-Item -Path "tests" -Destination "QGIS/src/plugins/primitive_editing/" -Recurse -Force
+New-Item -ItemType Directory -Force -Path $env:qgis_dir/src/plugins/primitive_editing
+Copy-Item -Path "$Env:PROJ_DIR/CMakeLists_inside.txt" -Destination "$env:qgis_dir/src/plugins/primitive_editing/CMakeLists.txt" -Force
+Copy-Item -Path "$Env:PROJ_DIR/src" -Destination "$env:qgis_dir/src/plugins/primitive_editing/" -Recurse -Force
+Copy-Item -Path "$Env:PROJ_DIR/resources" -Destination "$env:qgis_dir/src/plugins/primitive_editing/" -Recurse -Force
+Copy-Item -Path "$Env:PROJ_DIR/tests" -Destination "$env:qgis_dir/src/plugins/primitive_editing/" -Recurse -Force
 
 Write-Host "====== Prepare build..."
 
-# remove already installed python
-Remove-Item -Recurse C:/hostedtoolcache/windows/Python/
+if (Test-Path -Path C:/hostedtoolcache/windows/Python) {
+  # remove already installed python
+  Remove-Item -Recurse C:/hostedtoolcache/windows/Python/
+}
 
 # Update path to add Osgeo4W python support
 $Env:path="$Env:path;$Env:OSGEO4W_ROOT\apps\Python312;$Env:OSGEO4W_ROOT\apps\Python312\Scripts;$Env:OSGEO4W_ROOT\bin;"
@@ -25,14 +33,15 @@ $searchDir = "C:/Program Files (x86)/Windows Kits"
 $fileName = "SetupAPI.lib"
 Write-Host "Searching for $fileName in $searchDir..."
 
-$file = Get-ChildItem -Path $searchDir -Filter $fileName -Recurse -ErrorAction SilentlyContinue | Where-Object { $_.FullName -like "*x64*" } | Select-Object -First 1
+$file = Get-ChildItem -Path $searchDir -Filter $fileName -Recurse -ErrorAction Stop | Where-Object { $_.FullName -like "*x64*" } | Select-Object -First 1
 
 if ($file) {
-    Write-Host "File '$fileName' found at: $file"
-    $setupapidir = Split-Path $file -Parent
+  $full = $file | % { $_.FullName }
+  $setupapidir = Split-Path $full -Parent
+  Write-Host "File '$fileName' found $full in '$setupapidir'"
 } else {
-    Write-Error "File '$fileName' not found."
-    exit 1
+  Write-Error "File '$fileName' not found."
+  exit 1
 }
 
 # Disabled options:
@@ -44,7 +53,8 @@ Write-Host "====== Running cmake -S..."
 Write-Host "Checking for ZSTD libraries in $Env:OSGEO4W_ROOT\lib..."
 Get-ChildItem -Path "$Env:OSGEO4W_ROOT\lib" -Filter "zstd*"
 
-cmake -S $Env:PROJ_DIR/QGIS -B $Env:BUILD_DIR `
+cmake -S $env:qgis_dir -B $Env:BUILD_DIR `
+  -DCMAKE_CXX_FLAGS="/MP" `
   -DCMAKE_BUILD_TYPE=Release `
   -DAGGRESSIVE_SAFE_MODE=OFF `
   -DENABLE_MODELTEST=OFF `
@@ -80,9 +90,13 @@ cmake -S $Env:PROJ_DIR/QGIS -B $Env:BUILD_DIR `
   -DQt6_DIR="$Env:OSGEO4W_ROOT/apps/qt6/lib/cmake/Qt6" `
   -DCUSTOM_PLUGINS=primitive_editing
 
-
 Write-Host "====== Running cmake --target help..."
 cmake --build $Env:BUILD_DIR --config Release --target help
 
 Write-Host "====== Running cmake --build..."
-cmake --build $Env:BUILD_DIR --config Release --target plugin_primitiveediting
+cmake --build $Env:BUILD_DIR --config Release --target plugin_primitiveediting -j
+
+# dans CMakeList.txt
+# if(MSVC)
+# add_compile_options(/MP)
+# endif()
